refactor: extract mod `timer` into crate member
