+++ crates/timer/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+    name = "timer"
+    version = "0.1.0"
+    edition = "2021"
+
+[dependencies]
+    async-timers = "0.1.4"
+    tokio = {version = "1.39.2", features = [
+        "macros",
+        "sync",
+        "test-util",
+    ]}
