+++ src/raft/context.rs
@@ -2,11 +2,11 @@ use super::config::REQUEST_TIMEOUT;
 use super::log::LogManager;
 use super::service::PeerClient;
 use crate::conf::Config;
-use crate::timer::{OneshotTimer, PeriodicTimer};
 use log::debug;
 use std::collections::HashMap;
 use std::sync::Arc;
 use std::time::Duration;
+use timer::{OneshotTimer, PeriodicTimer};
 use tokio::sync::mpsc::{self, Sender};
 use tokio::sync::{Mutex, RwLock};
